Cache tests working
