+++ Makefile
@@ -3,7 +3,7 @@
 all: test flake
 
 test:
-	pytest --cov-report term-missing --cov=diffstream --verbose diffstream/tests/test_*py
+	pytest --cov-report term-missing --cov=cache --verbose tests/
 
 flake:
 	@echo 'flak8 output:'
@@ -12,3 +12,4 @@ flake:
 clean:
 	@echo "Removing .pyc cache files"
 	@find . -name "*.pyc" -delete
+	@find . -name "_pycache_" -rmdir
