+++ stream/consts.py
@@ -1,0 +1,9 @@
+"""
+Constants used for the communication stream
+"""
+
+PUB_SUB_PORT = 5563
+PUB_SUB_HOST = 'localhost'
+REQ_RES_PORT = 5564
+REQ_RES_HOST = 'localhost'
+TOPIC_STRING = '_ATOP_'
