+++ tests/test_protocol.py
@@ -1,5 +1,6 @@
-from diffstream import protocol
-from diffstream import consts
+from stream import protocol
+from cache import patch
+from cache import consts
 import uuid
